Bug-fix When making changes to a library and then changing the group and selecting another library, the original library was not being saved. This is because the library list was being cleared when the group selection was changed, and so an empty library would overwrite the library that needed to be saved.
To fix this, I have included a save routine in the group selection change handling code of the LibraryManagerWindow, and made it so that the LibraryBrowser notifies the action listeners of the group selection change before the LibraryBrowser repsonds to it. This means that the LibraryManagerWindow saves the library before the question list clears itself, etc.

Another important change is that the getSelectedLibraryID method now returns the pre-set selectedLibraryID variable, rather than calculating it again from the group name and selected index, etc. The selectedLibraryID is updated in the library selection change handling of the LibraryBrowser.

The selectedLibraryID must be set to null in the group selection change handling code of the LibraryBrowser to avoid the library selection change handling code of the LibraryManagerWindow re-saving the library that it points to (as the library is already saved by the group selection change handling code of the LibraryManagerWindow).