Bug fixes.
When removing an answer field, a ConcurrentModificationException was thrown because answer fields entries were being removed from the list that was being iterated. Now the entries are added to another array list 'forRemoval', and these are removed once the iteration is complete.

There were also a couple of other bugs that needed fixing: the RichTextArea is now able to handle the situation where a question contains an answer field that is no longer installed in ingatan. A couple of validation/repaint lines have also been added so that when an answer field is removed, the interface changes accordingly.