Altered the IAnswerField interface so that the quiz manager can set an actionlistener to any answer field added, if and only if it is the only answer field currently in the quiz answer area.
The idea is that the answer field can fire the listener upon an even that should logically result in a quiz continue action. This could be something like the user pressing enter in a text area. The answer field does not need to do anything with the listener if not required.

The reason the listener is only added if the answer field is the only one in the area is that we do not want any answer field to be able to push the quiz forward otherwise, because the user may not have yet entered data into other answer fields.