The ActionListener that the QuizWindow adds to answer fields is now added to all answer fields, not only if there is just one answer field in the current question. When an answer field fires the action listener (e.g. if the user presses enter in a text field, etc.), if there is only one field present for the current question then this is the equivalent of the user pressing the 'Continue' button. If there are multiple fields present, then the listener will focus the next answer field along.
An ArrayList of answer fields has been added to the QuizWindow which is populated each time a question is shown. The allows for easy access to the current answer fields in the answer text area, without having to traverse the text elements each time one requires access to a field.