Fix for issue ID = 1. Word wrap was failing for simple text answer fields embedded as part of a sentence. Words to the far right were clipped across the window frame before they were wrapped onto the new line. The text content of the answer fields was effecting wrapping (typing huge sentence into a fixed size field was causing the words to the right of the field to wrap down onto the next line, even though the field remained the same width.
The solution is to provide the JTextField with a columns value. In this case, the SimpleTextField used in AnsFieldSimplText.java is set to having 6 columns.

This seems to clear up the problem. The text now wraps too early (there is a small right margin), but this does not matter. Typing long sentences into a fixed width simple text answer field no longer results in text to the right of the field being wrapped to the next line.