Have completed work on the Embedded Audio answer field. It plays oggs using the JOgg and JOrbis libraries (www.jcraft.com) through the EasyOgg wrapper, written by Kevin from www.cokeandcode.com. I have added the extra licensing info to the website, the LGPL.TXT file included within the jar, and the About box.
The Embedded Audio answer field is three buttons: play/pause, stop, and load file. The load file disappears during quiz time. Pressing the play button will play the loaded file (if there is one) and change the icon to a pause icon. Pressing pause will change the button icon back to play, and pause playback. Pressing play after this will resume playback. Pressing stop at any time will close the file and reload it, and set the play button back to the play icon.

Also had to make it so that the FlexiQuestionContainer now calls its contextualiseAnswerFields methods on all three RichTextAreas, as answer fields can be added to any of the three areas now (if the answer field allows it). This change had to be made in the QuizWindow as well (not forgetting post-answer text).