Bug Fix. For table questions at quiz time with no question template, the question text was not showing because the setRichText method is used, and no [end] tag was added to the questionText string, unless a template was used.
Also, the text field was not being shown because it was being built before the data was set to it, and so it was not given a valid size. This has been fixed by firstly setting data to the field before it is built, and secondly making it so that it doesn't matter if the field has data when it is built :P If the field has no data, its default size is 60.

By 'built' I mean the field calls the 'rebuild()' method.