Made the answer field palette scroller a little wider and higher, to suit the size of the palette panel.
Fixed a bug with the edit answer fields button and the embedded audio answer field. The edit answer fields dialogue could not be shown because the embedded audio answer field instance was trying to call this.getParent().getParent(), which resulted in NullPointerException because in the case of the edit answer fields list, this.getParent() == null. Fixed this by testing whether this.getParent() == null in the embedded audio answer field.