Bug fix. Multiple selected lines (more correctly multiple paragraphs in the text component's document structure) were being copied with each paragraph being copied once for each paragraph selected. For example, "hello \n hello" is two paragraphs, and was being copied as "hello \n hello \n \n hello \n hello" or similar.
This was because the first and last runs within a paragraph were special cases, where not the entire run should be copied because the selection may not cover the entire run. Incorrect logic. It is only the first run of the first paragraph and the last run of the last paragraph that are special cases.

To fix this bug, the if statements have had (currentParagraph == startParagraph) and (currentParagraph == endParagraph) conditions added to them.