Have updated the simple text answer field edit context UI. It was bulky and superfluous, which does not make sense for a field often used inline with text. This change addresses Issue 13.
Along these lines also, the fill-in-the-list answer field has also received a UI update. The edit context was again, bulkier than required.

While updating the list UI, I found a bug in how the list is marked. The marker was not breaking up the possible answers for each entry before checking if an entry was correct or not.