Made it so that caret position in quiz time question and answer areas is set to 0 after question display.
Added a small work around to ensure the question and answer areas are being resized appropriately. Have resized the quiz window by -1 and 1 alternately each time a question is asked. Keeps the answer/question areas a nice size, and doesn't result in size change overall for the window. Will look into a better solution in future.