Carried out the Issue 26 task. When adding a non-answer field class as an answer field, a ClassDefNotFound error was being thrown. This error is now caught and a message is provided to the user.
Also new: if an exception or error occurs when trying to load the newly added answer field class, the copied class file will be deleted from the ingatan directory.... UNLESS... the user chose to delete the original class after copying (a message box allows them to choose this). If they did this, they are prompted before the copy is deleted.