This revision fixes issue 21 - where trying to draw a new atom and bond by dragging the atom edit tool would result in an exception.
The relevant section of code was found in the EditorPane class, in the dragComplete() method. Here, there is an if statement for the ATOM_TOOL tool (which is the identifier used for both the preset atoms tool *and* the atom label edit tool). This overriden method is called when the mouse is released by the CanvasMolecule super class if drag has occurred.

The new behaviour of the tool is that an atom labelled as "?" is added via a single bond as per the user's drag motion. The user can then click this atom immediately to edit its label. The reason that the text box doesn't come up automatically is that it is difficult to obtain the coordinates of the newly created atom. Mouse position at end of drag cannot be used as the bond is snapped to a grid, so the mouse location can be away from the terminating bond end. Also, the molecule object's atomX() and atomY() functions return relative grid spacings. See the pickAtom() method and the delcaration for px for hints on how to implement a function to get the location of an atom.