Adds functionality to the SketchEl chemistry sketching utility.
Users can now define their own SketchEl templates by drawing them and using the Save as Template menu item.

A new Ingatan "system" folder has been created (.ingatan/chem_templates). The path has been added to the IOManager class, so that there is now a 'getChemTemplatePath()' method. The IOManager also checks that the ./ingatan/chem_templates path exists on each load.

Any .el files that are saved to this new folder are loaded as templates, along with the templates bundled with Ingatan in the Ingatan executable .jar.

Custom templates must end with the extension .el or they will be ignored.