Implements Issue 29 - library manager should navigate to the most recently selected group automatically.
Have added a field to the preferences file `previouslySelectedGroup`. This has not required an upgrade to a newer preferences file type, as if the field is absent from the preferences file then the default value of "- All Libraries -" is taken, and the preferences file is updated on next close.

When the Library Manager loads, it sets the selected group to the one in the prefs file field (using IOManager.getPreviouslySelectedGroup() method). When the Library Manager closes, it updates the field with the selected group name using the appropriate IOManager method.