Updated the TableQuestionContainer UI so that the options pane has been replaced with a [Settings] button. When pressed, the option pane appears as a popup window. This required the font-name and question-ask-type combo boxes to be made into JSpinners, as the combo boxes would take focus from the popup and hence hide the popup. Also, the enterKeyMovesCellRight checkbox was removed from the option pane and added to the right of the [Settings] button.
A [Help] button has also been added to the QuestionTableContainer UI. This loads a dialog box that gives information on how to use the QuestionTable type.

The column names of the DataTable held by the QuestionTableContainer have been changed from "Questions" and "Answers" to "Side 1" and "Side 2" to re-inforce that this question type should be considered as being like flashcards.