Added OpenChart (approximatrix.com, LGPL) to the source package. This will be used to draw the graphs for the statistics centre. This revision includes OpenChart in the about window. The MainMenuWindow's Quiz History button now displays the StatsWindow. The QuizHistoryWindow has been converted to an extension of JPanel, and is now QuizHistoryPane.
Three more components have been added: GroupStatsGraphPane which shows statistics for a whole group, LibraryStatsGraphPane which shows graphical statistics for the selected libraries, and the RewardsPane which is a motivational device that can either be enjoyed/ignored by the user, where the user can use the points earnt doing quizzes to buy rewards. For example: "Bottle of Wine" for 4000 points. Or "Chocolate Bar" for 1000. There will be preset rewards, and these are customisable. Also, users may add new rewards with the graphic stored in the collections folder, with a new category automatically created called "rewards".